Updating the documentation using xlnt 1.x.x
This work is to test the new version of xlnt - released today. The installation is much simpler now due to the fact that xlnt no longer relies on third party libraries. This was a great move! I have confirmed that the minimum gcc version required is 6.2.0. work will still continue into exploring how the newest gcc 7.1 will impact xlnt. However, this will be covered under another heading and documentation issue/ticket and will involve installing gcc 7.1 from source and so forth. the documentation in this commit is purely intended to get a new users off and running with xlnt in the shortest possible time and with the least amount of effort.